Reorg NonEmptyList Instances
