--- kategory/src/main/kotlin/kategory/instances/NonEmptyListBimonad.kt
@@ -1,3 +1,0 @@
-package kategory
-
-interface NonEmptyListBimonad : Bimonad<NonEmptyList.F>, NonEmptyListMonad, NonEmptyListComonad
