--- kategory/src/main/kotlin/kategory/instances/NonEmptyListSemigroup.kt
@@ -1,5 +1,0 @@
-package kategory
-
-class NonEmptyListSemigroup<A> : Semigroup<NonEmptyList<A>> {
-    override fun combine(a: NonEmptyList<A>, b: NonEmptyList<A>): NonEmptyList<A> = a + b
-}
