+++ kategory/src/test/kotlin/kategory/data/NonEmptyListSemigroupTest.kt
@@ -9,7 +9,7 @@ class NonEmptyListSemigroupTest : UnitSpec() {
     init {
         "should semigroup with the instance passed" {
             forAll { value: Int ->
-                val nonEmptyListSemigroup = NonEmptyListSemigroup<Int>()
+                val nonEmptyListSemigroup = NonEmptyList.semigroup<Int>()
                 val seen = nonEmptyListSemigroup.combine(NonEmptyList.of(value), NonEmptyList.of(value))
                 val expected = NonEmptyList.of(value, value)
