+++ kategory/src/test/kotlin/kategory/data/OptionMonoidTest.kt
@@ -9,7 +9,7 @@ class OptionMonoidTest : UnitSpec() {
     init {
         "should semigroup with the instance passed" {
             forAll { _: Int ->
-                val optionMonoid = Option.monoid(NonEmptyListSemigroup<Int>())
+                val optionMonoid = Option.monoid(NonEmptyList.semigroup<Int>())
                 val seen = optionMonoid.combine(Option.None, Option.None)
                 val expected = Option.None
 
@@ -19,7 +19,7 @@ class OptionMonoidTest : UnitSpec() {
 
         "should semigroup with the instance passed" {
             forAll { value: Int ->
-                val optionMonoid = Option.monoid(NonEmptyListSemigroup<Int>())
+                val optionMonoid = Option.monoid(NonEmptyList.semigroup<Int>())
                 val seen = optionMonoid.combine(Option.Some(kategory.NonEmptyList.of(value)), Option.None)
                 val expected = Option.None
 
@@ -29,7 +29,7 @@ class OptionMonoidTest : UnitSpec() {
 
         "should semigroup with the instance passed" {
             forAll { value: Int ->
-                val optionMonoid = Option.monoid(NonEmptyListSemigroup<Int>())
+                val optionMonoid = Option.monoid(NonEmptyList.semigroup<Int>())
                 val seen = optionMonoid.combine(Option.None, Option.Some(NonEmptyList.of(value)))
                 val expected = Option.None
 
@@ -39,7 +39,7 @@ class OptionMonoidTest : UnitSpec() {
 
         "should semigroup with the instance passed" {
             forAll { value: Int ->
-                val optionMonoid = Option.monoid(NonEmptyListSemigroup<Int>())
+                val optionMonoid = Option.monoid(NonEmptyList.semigroup<Int>())
                 val seen = optionMonoid.combine(Option.Some(NonEmptyList.of(value)), Option.Some(NonEmptyList.of(value)))
                 val expected = Option.Some(NonEmptyList.of(value, value))
